requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
frozenlist==1.7.0
idna==3.10
imageio==2.37.0
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.5
mpmath==1.3.0
multidict==6.6.4
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
transformers==4.55.2
typing_extensions==4.14.1
yarl==1.20.1
ms-swift ==3.7.2
decord==0.6.0
torchvision==0.23.0


